+++ b2sdk/api.py
@@ -71,7 +71,7 @@ class Services(object):
         :param max_copy_workers: a number of copy threads
         :param max_download_workers: maximum number of download threads
         :param save_to_buffer_size: buffer size to use when writing files using DownloadedFile.save_to
-        :param check_download_hash: check hash of downloaded files
+        :param check_download_hash: whether to check hash of downloaded files. Can be disabled for files with internal checksums, for example, or to forcefully retrieve objects with corrupted payload or hash value
         """
         self.api = api
         self.session = api.session
@@ -142,7 +142,7 @@ class B2Api(metaclass=B2TraceMeta):
         :param api_config:
         :param max_download_workers: maximum number of download threads
         :param save_to_buffer_size: buffer size to use when writing files using DownloadedFile.save_to
-        :param check_download_hash: check hash of downloaded files
+        :param check_download_hash: whether to check hash of downloaded files. Can be disabled for files with internal checksums, for example, or to forcefully retrieve objects with corrupted payload or hash value
         """
         self.session = self.SESSION_CLASS(
             account_info=account_info, cache=cache, api_config=api_config