+++ b2sdk/transfer/inbound/downloader/parallel.py
@@ -117,7 +117,7 @@ class ParallelDownloader(AbstractDownloader):
         # Consume the rest of the file to complete the hashing process
         if self._check_hash:
             # we skip hashing if we would not check it - hasher object is actually a EmptyHasher instance
-            # but we avoid here reading whole file from disk again
+            # but we avoid here reading whole file (except for the first part) from disk again
             self._finish_hashing(first_part, file, hasher, download_version.content_length)
 
         return bytes_written, hasher.hexdigest()